feat: add catalog package
