+++ apps/cli/package.json
@@ -7,12 +7,14 @@
     "build": "bun build src/index.ts --outdir=dist --target=bun --minify",
     "build:types": "tsc --emitDeclarationOnly",
     "dev": "bun --watch run src/index.ts",
-    "start": "bun run src/index.ts init --dry-run",
+    "start": "bun run src/index.ts init -r ./.temp --dry-run",
+    "start:add": "bun run src/index.ts add -r ./.temp --dry-run",
     "type-check": "tsc --noEmit",
     "clean": "git clean -xdf .cache .turbo dist node_modules"
   },
   "dependencies": {
     "@effect/platform-bun": "4.0.0-beta.58",
+    "@repo/catalog": "workspace:*",
     "@repo/domain": "workspace:*",
     "@repo/scaffold": "workspace:*",
     "effect": "4.0.0-beta.58",
