+++ apps/cli/src/index.ts
@@ -1,4 +1,5 @@
 import { BunRuntime, BunServices } from "@effect/platform-bun";
+import { ModuleCatalog, TargetCatalog } from "@repo/catalog";
 import {
   ApplyService,
   BlueprintService,
@@ -17,6 +18,7 @@ const root = Command.make("stack-effect");
 const MainLayer = Layer.mergeAll(
   ApplyService.layer,
   BlueprintService.layer,
+  ModuleCatalog.layer,
   PlanService.layer,
   ScaffoldFormatter.layer,
   ConfigureService.layer,
