+++ packages/catalog/src/index.ts
@@ -1,0 +1,2 @@
+export { ModuleCatalog } from "./ModuleCatalog";
+export { TargetCatalog } from "./TargetCatalog";
