+++ packages/catalog/src/registry/content/api.ts
