+++ packages/catalog/src/registry/content/init.ts
