+++ packages/catalog/src/registry/content/server.ts
