+++ packages/catalog/src/registry/content/shared.ts
