+++ packages/catalog/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "@repo/config-typescript/base.json",
+  "compilerOptions": {
+    "rootDir": ".",
+    "lib": ["ES2023", "DOM"],
+    "composite": true,
+    "declaration": true,
+    "declarationMap": true,
+    "outDir": "./dist",
+    "skipLibCheck": true,
+    "types": ["node"]
+  },
+  "include": ["src", "scratchpad.ts"],
+  "exclude": ["dist", "build", "node_modules"]
+}
